lduy chinh bill detail va employee

+++ Entity/BillDetail.cs
@@ -13,12 +13,11 @@ namespace Entity
         [Key]
         public int Bill_Detail_ID { get; set; }
         [ForeignKey("Book")]
-
         public int Book_ID { get; set; }
         [ForeignKey("Bill")]
         public int Bill_ID { get; set; }
-        public int Amount { get; set; }
-        public float Price { get; set; }
-
+        public decimal Amount { get; set; }
+        public decimal Price { get; set; }
+       
     }
 }
